Test needing chgroups will indeterminate if unsupoorted
If we know that we're running on a key that can't handle channel group
API, tests that need said API will report their result as
"indeterminate". IOW, they won't fail, but won't pretend to
succeed/pass either.

This, as of now, "automagically" detected on Travis builds of Pull
Requests, as it doesn't pass the encrypted environment variables to
such builds (for security reasons).

We might defer this to a command-line argument, rather than detecting
it ourelves, this was quick way to do it.