Deem test indtereminate on exceed of quota
In general, exceeding quota simply means tests are being run on a
restricted account. The C-core is not itself at fault here, yet,
test can't be correctly run.

Of course, for some special test it doesn't have to be like that, but,
such test can use its own "check infrastructure", instead of the
general one.